WIP implementation of react-navigation for home screen
