+++ App.js
@@ -1,24 +1,19 @@
-import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
+import { createSwitchNavigator, createAppContainer } from "react-navigation";
+import { useScreens } from 'react-native-screens';
+useScreens();
 
+import HomeScreen from './HomeScreen';
 import SessionPlayer from './SessionPlayer';
 
-export default class App extends React.Component {
-    render() {
-        return (
-            <View style={styles.container}>
-                <SessionPlayer />
-            </View>
-        );
-    }
-}
-
-const styles = StyleSheet.create({
-    container: {
-        width: '100%',
-        flex: 1,
-        backgroundColor: '#fff',
-        alignItems: 'center',
-        justifyContent: 'center',
+const AppNavigator = createSwitchNavigator({
+    HomeScreen: {
+        screen: HomeScreen,
     },
+    SessionPlayer: {
+        screen: SessionPlayer,
+    },
+}, {
+    initialRouteName: 'HomeScreen',
 });
+
+export default createAppContainer(AppNavigator);
