+++ HomeScreen/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "HomeScreen",
+  "main": "./HomeScreen.js"
+}
+
